Revert "8-byte 'long' is replaced with 4-byte 'int'"
This reverts commit bb73f5040c01f47b1bf98f2feab5bca523a74881.

+++ src/general/cell.h
@@ -30,7 +30,7 @@ protected:
   double measure_value;
 
 public:
-  int cellno,nodeno[4],faceno[4],neighbour_cellno[4];
+  long int cellno,nodeno[4],faceno[4],neighbour_cellno[4];
   NodeType* node[4];
   FaceType* face[4];
